refactor(test): rename AggregateExpecter and extend its functionality
- Rename AggregateExpecter to ExpectDsl
- Move expect function to DefaultExpectDsl class
- Update ExpectStage interface to extend AggregateExpecter
- Refactor test DSL to use new ExpectDsl interface
