+++ test/wow-test/src/main/kotlin/me/ahoo/wow/test/aggregate/ExpectStage.kt
@@ -13,165 +13,15 @@
 
 package me.ahoo.wow.test.aggregate
 
-import me.ahoo.test.asserts.assert
-import me.ahoo.wow.api.event.DomainEvent
-import me.ahoo.wow.event.DomainEventStream
 import me.ahoo.wow.ioc.ServiceProvider
 import me.ahoo.wow.modeling.command.CommandAggregateFactory
 import me.ahoo.wow.modeling.matedata.AggregateMetadata
-import me.ahoo.wow.modeling.state.StateAggregate
 import org.assertj.core.error.MultipleAssertionsError
 import reactor.core.publisher.Mono
 import reactor.kotlin.test.test
 import java.util.function.Consumer
-import kotlin.reflect.KClass
 
-interface ExpectStage<S : Any> {
-    fun expect(expected: ExpectedResult<S>.() -> Unit): ExpectStage<S>
-
-    /**
-     * 3.1 期望聚合状态.
-     */
-    fun expectStateAggregate(expected: StateAggregate<S>.() -> Unit): ExpectStage<S> {
-        return expect { expected(stateAggregate) }
-    }
-
-    fun expectState(expected: S.() -> Unit): ExpectStage<S> {
-        return expectStateAggregate { expected(state) }
-    }
-
-    fun expectState(expected: Consumer<S>): ExpectStage<S> {
-        return expectState { expected.accept(this) }
-    }
-
-    /**
-     * 3.2 期望领域事件.
-     */
-    fun expectEventStream(expected: DomainEventStream.() -> Unit): ExpectStage<S> {
-        return expect {
-            domainEventStream.assert().describedAs {
-                buildString {
-                    append("Expect the domain event stream is not null.")
-                    error?.let { error ->
-                        appendLine()
-                        append(error.stackTraceToString())
-                    }
-                }
-            }.isNotNull()
-            expected(domainEventStream!!)
-        }
-    }
-
-    fun expectEventStream(expected: Consumer<DomainEventStream>): ExpectStage<S> {
-        return expectEventStream {
-            expected.accept(this)
-        }
-    }
-
-    fun expectEventIterator(expected: EventIterator.() -> Unit): ExpectStage<S> {
-        return expectEventStream {
-            expected(EventIterator((iterator())))
-        }
-    }
-
-    fun expectEventIterator(expected: Consumer<EventIterator>): ExpectStage<S> {
-        return expectEventIterator {
-            expected.accept(this)
-        }
-    }
-
-    /**
-     * 期望的第一个领域事件
-     */
-    fun <E : Any> expectEvent(expected: DomainEvent<E>.() -> Unit): ExpectStage<S> {
-        return expectEventStream {
-            assert().describedAs { "Expect the domain event stream size to be greater than 1." }
-                .hasSizeGreaterThanOrEqualTo(1)
-            @Suppress("UNCHECKED_CAST")
-            expected(first() as DomainEvent<E>)
-        }
-    }
-
-    fun <E : Any> expectEvent(expected: Consumer<DomainEvent<E>>): ExpectStage<S> {
-        return expectEvent {
-            expected.accept(this)
-        }
-    }
-
-    fun <E : Any> expectEventBody(expected: E.() -> Unit): ExpectStage<S> {
-        return expectEvent {
-            expected(body)
-        }
-    }
-
-    fun <E : Any> expectEventBody(expected: Consumer<E>): ExpectStage<S> {
-        return expectEventBody<E> {
-            expected.accept(this)
-        }
-    }
-
-    /**
-     * 期望产生的事件数量.
-     */
-    fun expectEventCount(expected: Int): ExpectStage<S> {
-        return expectEventStream {
-            assert().describedAs { "Expect the domain event stream size." }.hasSize(expected)
-        }
-    }
-
-    fun expectEventType(vararg expected: KClass<*>): ExpectStage<S> {
-        return expectEventType(*expected.map { it.java }.toTypedArray())
-    }
-
-    /**
-     * 期望事件类型. 严格按照事件生成顺序验证。
-     */
-    fun expectEventType(vararg expected: Class<*>): ExpectStage<S> {
-        return expectEventCount(expected.size).expectEventStream {
-            val itr = iterator()
-            for (eventType in expected) {
-                itr.next().body.assert().isInstanceOf(eventType)
-            }
-        }
-    }
-
-    fun expectNoError(): ExpectStage<S> {
-        return expect {
-            error.assert().describedAs { "Expect no error" }.isNull()
-        }
-    }
-
-    fun expectError(): ExpectStage<S> {
-        return expect {
-            error.assert().describedAs { "Expect an error." }.isNotNull()
-        }
-    }
-
-    /**
-     * 3.3 期望产生异常
-     */
-    fun <E : Throwable> expectError(expected: E.() -> Unit): ExpectStage<S> {
-        return expectError().expect {
-            @Suppress("UNCHECKED_CAST")
-            expected(error as E)
-        }
-    }
-
-    fun <E : Throwable> expectError(expected: Consumer<E>): ExpectStage<S> {
-        return expectError<E> {
-            expected.accept(this)
-        }
-    }
-
-    fun <E : Throwable> expectErrorType(expected: KClass<E>): ExpectStage<S> {
-        return expectErrorType(expected.java)
-    }
-
-    fun <E : Throwable> expectErrorType(expected: Class<E>): ExpectStage<S> {
-        return expectError<E> {
-            assert().isInstanceOf(expected)
-        }
-    }
+interface ExpectStage<S : Any> : AggregateExpecter<S, ExpectStage<S>> {
     fun verify(): VerifiedStage<S> {
         return verify(true)
     }
